+++ src/lib.rs
@@ -9,6 +9,9 @@ pub mod error;
 pub mod s3;
 pub mod terraform;
 
+#[cfg(test)]
+mod tests;
+
 use crate::ansible::{AnsibleRunner, AnsibleRunnerInterface};
 use crate::error::{Error, Result};
 use crate::s3::S3AssetRepository;
@@ -41,7 +44,6 @@ pub struct TestnetDeployBuilder {
     provider: Option<CloudProvider>,
     state_bucket_name: Option<String>,
     terraform_binary_path: Option<PathBuf>,
-    ansible_binary_path: Option<PathBuf>,
     working_directory_path: Option<PathBuf>,
     ssh_secret_key_path: Option<PathBuf>,
     vault_password_path: Option<PathBuf>,
@@ -365,660 +367,4 @@ pub fn run_external_command(
     }
 
     Ok(output_lines)
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::ansible::MockAnsibleRunnerInterface;
-    use crate::s3::S3AssetRepository;
-    use crate::terraform::MockTerraformRunnerInterface;
-    use assert_fs::fixture::ChildPath;
-    use assert_fs::prelude::*;
-    use assert_fs::TempDir;
-    use color_eyre::{eyre::eyre, Result};
-    use flate2::write::GzEncoder;
-    use flate2::Compression;
-    use httpmock::prelude::*;
-    use mockall::predicate::*;
-    use mockall::Sequence;
-    use std::fs::{File, Metadata};
-    use std::os::unix::fs::PermissionsExt;
-
-    const RPC_CLIENT_BIN_NAME: &str = "safenode_rpc_client";
-
-    fn setup_working_directory() -> Result<(TempDir, ChildPath)> {
-        let tmp_dir = assert_fs::TempDir::new()?;
-        let working_dir = tmp_dir.child("work");
-        working_dir.create_dir_all()?;
-        working_dir.copy_from("resources/", &["**"])?;
-        Ok((tmp_dir, working_dir))
-    }
-
-    fn create_fake_rpc_client_archive(working_dir: &ChildPath) -> Result<(ChildPath, Metadata)> {
-        let temp_archive_dir = working_dir.child("setup_archive");
-        let rpc_client_archive = temp_archive_dir.child("rpc_client.tar.gz");
-        let fake_rpc_client_bin = temp_archive_dir.child(RPC_CLIENT_BIN_NAME);
-        fake_rpc_client_bin.write_binary(b"fake code")?;
-
-        let mut fake_rpc_client_bin_file = File::open(fake_rpc_client_bin.path())?;
-        let gz_encoder = GzEncoder::new(
-            File::create(rpc_client_archive.path())?,
-            Compression::default(),
-        );
-        let mut builder = tar::Builder::new(gz_encoder);
-        builder.append_file(RPC_CLIENT_BIN_NAME, &mut fake_rpc_client_bin_file)?;
-        builder.into_inner()?;
-        let rpc_client_archive_metadata = std::fs::metadata(rpc_client_archive.path())?;
-
-        Ok((rpc_client_archive, rpc_client_archive_metadata))
-    }
-
-    fn setup_default_ansible_runner(name: &str) -> MockAnsibleRunnerInterface {
-        let mut ansible_runner = MockAnsibleRunnerInterface::new();
-        ansible_runner
-            .expect_run_playbook()
-            .times(1)
-            .with(
-                eq(PathBuf::from("genesis_node.yml")),
-                eq(PathBuf::from("inventory")
-                    .join(format!(".{name}_genesis_inventory_digital_ocean.yml"))),
-                eq("root".to_string()),
-                eq(Some(
-                    "{ \"is_genesis\": \"True\", \"provider\": \"digital-ocean\" }".to_string(),
-                )),
-            )
-            .returning(|_, _, _, _| Ok(()));
-        ansible_runner
-    }
-
-    fn setup_default_terraform_runner(name: &str) -> MockTerraformRunnerInterface {
-        let mut terraform_runner = MockTerraformRunnerInterface::new();
-        terraform_runner.expect_init().times(1).returning(|| Ok(()));
-        terraform_runner
-            .expect_workspace_list()
-            .times(1)
-            .returning(|| Ok(vec!["default".to_string(), "dev".to_string()]));
-        terraform_runner
-            .expect_workspace_new()
-            .times(1)
-            .with(eq(name.to_string()))
-            .returning(|_| Ok(()));
-        terraform_runner
-    }
-
-    fn setup_default_s3_repository(working_dir: &ChildPath) -> Result<S3AssetRepository> {
-        let (rpc_client_archive, rpc_client_archive_metadata) =
-            create_fake_rpc_client_archive(working_dir)?;
-        let asset_server = MockServer::start();
-        asset_server.mock(|when, then| {
-            when.method(GET)
-                .path("/rpc_client-latest-x86_64-unknown-linux-musl.tar.gz");
-            then.status(200)
-                .header(
-                    "Content-Length",
-                    rpc_client_archive_metadata.len().to_string(),
-                )
-                .header("Content-Type", "application/gzip")
-                .body_from_file(rpc_client_archive.path().to_str().unwrap());
-        });
-
-        let s3_repository = S3AssetRepository::new(&asset_server.base_url());
-        Ok(s3_repository)
-    }
-
-    #[tokio::test]
-    async fn init_should_create_a_new_workspace() -> Result<()> {
-        let (tmp_dir, working_dir) = setup_working_directory()?;
-        let mut terraform_runner = MockTerraformRunnerInterface::new();
-        terraform_runner.expect_init().times(1).returning(|| Ok(()));
-        terraform_runner
-            .expect_workspace_list()
-            .times(1)
-            .returning(|| Ok(vec!["default".to_string(), "dev".to_string()]));
-        terraform_runner
-            .expect_workspace_new()
-            .times(1)
-            .with(eq("beta".to_string()))
-            .returning(|_| Ok(()));
-        let s3_repository = setup_default_s3_repository(&working_dir)?;
-        let testnet = TestnetDeploy::new(
-            Box::new(terraform_runner),
-            Box::new(MockAnsibleRunnerInterface::new()),
-            working_dir.to_path_buf(),
-            CloudProvider::DigitalOcean,
-            s3_repository,
-        );
-        testnet.init("beta").await?;
-        drop(tmp_dir);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn init_should_not_create_a_new_workspace_when_one_with_the_same_name_exists(
-    ) -> Result<()> {
-        let (tmp_dir, working_dir) = setup_working_directory()?;
-        let mut terraform_runner = MockTerraformRunnerInterface::new();
-        terraform_runner.expect_init().times(1).returning(|| Ok(()));
-        terraform_runner
-            .expect_workspace_list()
-            .times(1)
-            .returning(|| {
-                Ok(vec![
-                    "alpha".to_string(),
-                    "default".to_string(),
-                    "dev".to_string(),
-                ])
-            });
-        terraform_runner
-            .expect_workspace_new()
-            .times(0)
-            .with(eq("alpha".to_string()))
-            .returning(|_| Ok(()));
-
-        let s3_repository = setup_default_s3_repository(&working_dir)?;
-        let testnet = TestnetDeploy::new(
-            Box::new(terraform_runner),
-            Box::new(MockAnsibleRunnerInterface::new()),
-            working_dir.to_path_buf(),
-            CloudProvider::DigitalOcean,
-            s3_repository,
-        );
-        testnet.init("alpha").await?;
-        drop(tmp_dir);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn init_should_download_and_extract_the_rpc_client() -> Result<()> {
-        let (tmp_dir, working_dir) = setup_working_directory()?;
-        let temp_archive_dir = working_dir.child("setup_archive");
-
-        // Create an archive containing a fake rpc client exe, to be returned by the mock HTTP
-        // server.
-        let rpc_client_archive = temp_archive_dir.child("rpc_client.tar.gz");
-        let fake_rpc_client_bin = temp_archive_dir.child(RPC_CLIENT_BIN_NAME);
-        fake_rpc_client_bin.write_binary(b"fake code")?;
-        let mut fake_rpc_client_bin_file = File::open(fake_rpc_client_bin.path())?;
-        let gz_encoder = GzEncoder::new(
-            File::create(rpc_client_archive.path())?,
-            Compression::default(),
-        );
-        let mut builder = tar::Builder::new(gz_encoder);
-        builder.append_file(RPC_CLIENT_BIN_NAME, &mut fake_rpc_client_bin_file)?;
-        builder.into_inner()?;
-        let rpc_client_archive_metadata = std::fs::metadata(rpc_client_archive.path())?;
-
-        let asset_server = MockServer::start();
-        asset_server.mock(|when, then| {
-            when.method(GET)
-                .path("/rpc_client-latest-x86_64-unknown-linux-musl.tar.gz");
-            then.status(200)
-                .header(
-                    "Content-Length",
-                    rpc_client_archive_metadata.len().to_string(),
-                )
-                .header("Content-Type", "application/gzip")
-                .body_from_file(rpc_client_archive.path().to_str().unwrap());
-        });
-        let downloaded_safe_archive =
-            working_dir.child("rpc_client-latest-x86_64-unknown-linux-musl.tar.gz");
-
-        let extracted_rpc_client_bin = working_dir.child(RPC_CLIENT_BIN_NAME);
-        let s3_repository = S3AssetRepository::new(&asset_server.base_url());
-        let terraform_runner = setup_default_terraform_runner("alpha");
-        let testnet = TestnetDeploy::new(
-            Box::new(terraform_runner),
-            Box::new(MockAnsibleRunnerInterface::new()),
-            working_dir.to_path_buf(),
-            CloudProvider::DigitalOcean,
-            s3_repository,
-        );
-
-        testnet.init("alpha").await?;
-
-        downloaded_safe_archive.assert(predicates::path::missing());
-        extracted_rpc_client_bin.assert(predicates::path::is_file());
-
-        let metadata = std::fs::metadata(extracted_rpc_client_bin.path())?;
-        let permissions = metadata.permissions();
-        assert!(permissions.mode() & 0o100 > 0, "File is not executable");
-        drop(tmp_dir);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn init_should_not_download_the_rpc_client_if_it_already_exists() -> Result<()> {
-        let (tmp_dir, working_dir) = setup_working_directory()?;
-        let fake_rpc_client_bin = working_dir.child(RPC_CLIENT_BIN_NAME);
-        fake_rpc_client_bin.write_binary(b"fake code")?;
-
-        let (rpc_client_archive, rpc_client_archive_metadata) =
-            create_fake_rpc_client_archive(&working_dir)?;
-        let asset_server = MockServer::start();
-        let mock = asset_server.mock(|when, then| {
-            when.method(GET)
-                .path("/rpc_client-latest-x86_64-unknown-linux-musl.tar.gz");
-            then.status(200)
-                .header(
-                    "Content-Length",
-                    rpc_client_archive_metadata.len().to_string(),
-                )
-                .header("Content-Type", "application/gzip")
-                .body_from_file(rpc_client_archive.path().to_str().unwrap());
-        });
-        let s3_repository = S3AssetRepository::new(&asset_server.base_url());
-
-        let terraform_runner = setup_default_terraform_runner("alpha");
-        let testnet = TestnetDeploy::new(
-            Box::new(terraform_runner),
-            Box::new(MockAnsibleRunnerInterface::new()),
-            working_dir.to_path_buf(),
-            CloudProvider::DigitalOcean,
-            s3_repository,
-        );
-        testnet.init("alpha").await?;
-
-        mock.assert_hits(0);
-        drop(tmp_dir);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn init_should_generate_ansible_inventory_for_digital_ocean_for_the_new_testnet(
-    ) -> Result<()> {
-        let (tmp_dir, working_dir) = setup_working_directory()?;
-        let s3_repository = setup_default_s3_repository(&working_dir)?;
-        let terraform_runner = setup_default_terraform_runner("alpha");
-
-        let testnet = TestnetDeploy::new(
-            Box::new(terraform_runner),
-            Box::new(MockAnsibleRunnerInterface::new()),
-            working_dir.to_path_buf(),
-            CloudProvider::DigitalOcean,
-            s3_repository,
-        );
-
-        testnet.init("alpha").await?;
-
-        let inventory_files = ["build", "genesis", "node"];
-        for inventory_type in inventory_files.iter() {
-            let inventory_file = working_dir.child(format!(
-                "ansible/inventory/.{}_{}_inventory_digital_ocean.yml",
-                "alpha", inventory_type
-            ));
-            inventory_file.assert(predicates::path::is_file());
-
-            let contents = std::fs::read_to_string(inventory_file.path())?;
-            assert!(contents.contains("alpha"));
-            assert!(contents.contains(inventory_type));
-        }
-        drop(tmp_dir);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn init_should_not_overwrite_generated_inventory() -> Result<()> {
-        let (tmp_dir, working_dir) = setup_working_directory()?;
-        let s3_repository = setup_default_s3_repository(&working_dir)?;
-        let mut terraform_runner = MockTerraformRunnerInterface::new();
-        let mut seq = Sequence::new();
-
-        terraform_runner.expect_init().times(2).returning(|| Ok(()));
-        terraform_runner
-            .expect_workspace_list()
-            .times(1)
-            .in_sequence(&mut seq)
-            .returning(|| Ok(vec!["default".to_string(), "dev".to_string()]));
-        terraform_runner
-            .expect_workspace_new()
-            .times(1)
-            .with(eq("alpha".to_string()))
-            .returning(|_| Ok(()));
-        terraform_runner
-            .expect_workspace_list()
-            .times(1)
-            .in_sequence(&mut seq)
-            .returning(|| {
-                Ok(vec![
-                    "alpha".to_string(),
-                    "default".to_string(),
-                    "dev".to_string(),
-                ])
-            });
-
-        let testnet = TestnetDeploy::new(
-            Box::new(terraform_runner),
-            Box::new(MockAnsibleRunnerInterface::new()),
-            working_dir.to_path_buf(),
-            CloudProvider::DigitalOcean,
-            s3_repository,
-        );
-
-        testnet.init("alpha").await?;
-        testnet.init("alpha").await?; // this should be idempotent
-
-        let inventory_files = ["build", "genesis", "node"];
-        for inventory_type in inventory_files.iter() {
-            let inventory_file = working_dir.child(format!(
-                "ansible/inventory/.{}_{}_inventory_digital_ocean.yml",
-                "alpha", inventory_type
-            ));
-            inventory_file.assert(predicates::path::is_file());
-
-            let contents = std::fs::read_to_string(inventory_file.path())?;
-            assert!(contents.contains("alpha"));
-            assert!(contents.contains(inventory_type));
-        }
-        drop(tmp_dir);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn deploy_should_run_terraform_apply() -> Result<()> {
-        let (tmp_dir, working_dir) = setup_working_directory()?;
-        let s3_repository = setup_default_s3_repository(&working_dir)?;
-        let mut terraform_runner = MockTerraformRunnerInterface::new();
-        terraform_runner
-            .expect_workspace_select()
-            .times(1)
-            .with(eq("beta".to_string()))
-            .returning(|_| Ok(()));
-        terraform_runner
-            .expect_apply()
-            .times(1)
-            .with(eq(vec![
-                ("node_count".to_string(), "30".to_string()),
-                ("use_custom_bin".to_string(), "false".to_string()),
-            ]))
-            .returning(|_| Ok(()));
-
-        let testnet = TestnetDeploy::new(
-            Box::new(terraform_runner),
-            Box::new(setup_default_ansible_runner("beta")),
-            working_dir.to_path_buf(),
-            CloudProvider::DigitalOcean,
-            s3_repository,
-        );
-
-        testnet.deploy("beta", 30, None, None).await?;
-
-        drop(tmp_dir);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn deploy_should_run_terraform_apply_with_custom_bin_set_when_repo_is_supplied(
-    ) -> Result<()> {
-        let (tmp_dir, working_dir) = setup_working_directory()?;
-        let s3_repository = setup_default_s3_repository(&working_dir)?;
-        let mut terraform_runner = MockTerraformRunnerInterface::new();
-        terraform_runner
-            .expect_workspace_select()
-            .times(1)
-            .with(eq("beta".to_string()))
-            .returning(|_| Ok(()));
-        terraform_runner
-            .expect_apply()
-            .times(1)
-            .with(eq(vec![
-                ("node_count".to_string(), "30".to_string()),
-                ("use_custom_bin".to_string(), "true".to_string()),
-            ]))
-            .returning(|_| Ok(()));
-
-        let testnet = TestnetDeploy::new(
-            Box::new(terraform_runner),
-            Box::new(setup_default_ansible_runner("beta")),
-            working_dir.to_path_buf(),
-            CloudProvider::DigitalOcean,
-            s3_repository,
-        );
-
-        testnet
-            .deploy(
-                "beta",
-                30,
-                Some("maidsafe".to_string()),
-                Some("custom_branch".to_string()),
-            )
-            .await?;
-
-        drop(tmp_dir);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn deploy_should_ensure_branch_is_set_if_repo_owner_is_used() -> Result<()> {
-        let (tmp_dir, working_dir) = setup_working_directory()?;
-        let s3_repository = setup_default_s3_repository(&working_dir)?;
-        let mut terraform_runner = MockTerraformRunnerInterface::new();
-        terraform_runner
-            .expect_workspace_select()
-            .times(0)
-            .returning(|_| Ok(()));
-        terraform_runner
-            .expect_apply()
-            .times(0)
-            .returning(|_| Ok(()));
-
-        let testnet = TestnetDeploy::new(
-            Box::new(terraform_runner),
-            Box::new(MockAnsibleRunnerInterface::new()),
-            working_dir.to_path_buf(),
-            CloudProvider::DigitalOcean,
-            s3_repository,
-        );
-
-        let result = testnet
-            .deploy("beta", 30, Some("maidsafe".to_string()), None)
-            .await;
-
-        match result {
-            Ok(()) => {
-                drop(tmp_dir);
-                Err(eyre!("deploy should have returned an error"))
-            }
-            Err(e) => {
-                assert_eq!(
-                    e.to_string(),
-                    "Both the repository owner and branch name must be supplied if either are used"
-                );
-                drop(tmp_dir);
-                Ok(())
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn deploy_should_ensure_repo_owner_is_set_if_branch_is_used() -> Result<()> {
-        let (tmp_dir, working_dir) = setup_working_directory()?;
-        let s3_repository = setup_default_s3_repository(&working_dir)?;
-        let mut terraform_runner = MockTerraformRunnerInterface::new();
-        terraform_runner
-            .expect_workspace_select()
-            .times(0)
-            .returning(|_| Ok(()));
-        terraform_runner
-            .expect_apply()
-            .times(0)
-            .returning(|_| Ok(()));
-
-        let testnet = TestnetDeploy::new(
-            Box::new(terraform_runner),
-            Box::new(MockAnsibleRunnerInterface::new()),
-            working_dir.to_path_buf(),
-            CloudProvider::DigitalOcean,
-            s3_repository,
-        );
-
-        let result = testnet
-            .deploy("beta", 30, None, Some("custom_branch".to_string()))
-            .await;
-
-        match result {
-            Ok(()) => {
-                drop(tmp_dir);
-                Err(eyre!("deploy should have returned an error"))
-            }
-            Err(e) => {
-                assert_eq!(
-                    e.to_string(),
-                    "Both the repository owner and branch name must be supplied if either are used"
-                );
-                drop(tmp_dir);
-                Ok(())
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn deploy_should_run_ansible_against_genesis_then_remaining_nodes() -> Result<()> {
-        let (tmp_dir, working_dir) = setup_working_directory()?;
-        let s3_repository = setup_default_s3_repository(&working_dir)?;
-        let mut terraform_runner = setup_default_terraform_runner("beta");
-        terraform_runner
-            .expect_workspace_select()
-            .times(1)
-            .with(eq("beta".to_string()))
-            .returning(|_| Ok(()));
-        terraform_runner
-            .expect_apply()
-            .times(1)
-            .with(eq(vec![
-                ("node_count".to_string(), "30".to_string()),
-                ("use_custom_bin".to_string(), "false".to_string()),
-            ]))
-            .returning(|_| Ok(()));
-        let mut ansible_runner = MockAnsibleRunnerInterface::new();
-        ansible_runner
-            .expect_run_playbook()
-            .times(1)
-            .with(
-                eq(PathBuf::from("genesis_node.yml")),
-                eq(PathBuf::from("inventory").join(".beta_genesis_inventory_digital_ocean.yml")),
-                eq("root".to_string()),
-                eq(Some(
-                    "{ \"is_genesis\": \"true\", \"provider\": \"digital-ocean\" }".to_string(),
-                )),
-            )
-            .returning(|_, _, _, _| Ok(()));
-
-        let testnet = TestnetDeploy::new(
-            Box::new(terraform_runner),
-            Box::new(ansible_runner),
-            working_dir.to_path_buf(),
-            CloudProvider::DigitalOcean,
-            s3_repository,
-        );
-
-        testnet.init("beta").await?;
-        testnet.deploy("beta", 30, None, None).await?;
-
-        drop(tmp_dir);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn clean_should_run_terraform_destroy_and_delete_workspace_and_delete_inventory_files(
-    ) -> Result<()> {
-        let (tmp_dir, working_dir) = setup_working_directory()?;
-        let s3_repository = setup_default_s3_repository(&working_dir)?;
-        let mut terraform_runner = setup_default_terraform_runner("alpha");
-        let mut seq = Sequence::new();
-        terraform_runner
-            .expect_workspace_list()
-            .times(1)
-            .returning(|| {
-                Ok(vec![
-                    "alpha".to_string(),
-                    "default".to_string(),
-                    "dev".to_string(),
-                ])
-            });
-        terraform_runner
-            .expect_workspace_select()
-            .times(1)
-            .in_sequence(&mut seq)
-            .with(eq("alpha".to_string()))
-            .returning(|_| Ok(()));
-        terraform_runner
-            .expect_destroy()
-            .times(1)
-            .returning(|| Ok(()));
-        terraform_runner
-            .expect_workspace_select()
-            .times(1)
-            .in_sequence(&mut seq)
-            .with(eq("dev".to_string()))
-            .returning(|_| Ok(()));
-        terraform_runner
-            .expect_workspace_delete()
-            .times(1)
-            .with(eq("alpha".to_string()))
-            .returning(|_| Ok(()));
-
-        let testnet = TestnetDeploy::new(
-            Box::new(terraform_runner),
-            Box::new(MockAnsibleRunnerInterface::new()),
-            working_dir.to_path_buf(),
-            CloudProvider::DigitalOcean,
-            s3_repository,
-        );
-
-        // Calling init will create the Ansible inventory files, which we want to be removed by
-        // the clean operation.
-        testnet.init("alpha").await?;
-        testnet.clean("alpha").await?;
-
-        let inventory_types = ["build", "genesis", "node"];
-        for inventory_type in inventory_types.iter() {
-            let inventory_file = working_dir.child(format!(
-                "ansible/inventory/.{}_{}_inventory_digital_ocean.yml",
-                "alpha", inventory_type
-            ));
-            inventory_file.assert(predicates::path::missing());
-        }
-
-        drop(tmp_dir);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn clean_with_invalid_name_should_return_an_error() -> Result<()> {
-        let (tmp_dir, working_dir) = setup_working_directory()?;
-        let s3_repository = setup_default_s3_repository(&working_dir)?;
-        let mut terraform_runner = MockTerraformRunnerInterface::new();
-        terraform_runner
-            .expect_workspace_list()
-            .times(1)
-            .returning(|| {
-                Ok(vec![
-                    "alpha".to_string(),
-                    "default".to_string(),
-                    "dev".to_string(),
-                ])
-            });
-
-        let testnet = TestnetDeploy::new(
-            Box::new(terraform_runner),
-            Box::new(MockAnsibleRunnerInterface::new()),
-            working_dir.to_path_buf(),
-            CloudProvider::DigitalOcean,
-            s3_repository,
-        );
-
-        let result = testnet.clean("beta").await;
-        match result {
-            Ok(()) => {
-                drop(tmp_dir);
-                Err(eyre!("deploy should have returned an error"))
-            }
-            Err(e) => {
-                assert_eq!(e.to_string(), "The 'beta' environment does not exist");
-                drop(tmp_dir);
-                Ok(())
-            }
-        }
-    }
 }
