+++ src/tests/mod.rs
@@ -1,0 +1,6 @@
+mod clean;
+mod deploy;
+mod init;
+mod setup;
+
+const RPC_CLIENT_BIN_NAME: &str = "safenode_rpc_client";
